ResumeService/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
asgiref==3.9.1
av==15.1.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
decorator==4.4.2
Django==5.2.6
djangorestframework==3.16.1
dnspython==2.7.0
dotenv==0.9.9
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-speech==2.33.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
lxml==6.0.1
mongoengine==0.29.1
moviepy==1.0.3
mpmath==1.3.0
numpy==2.3.2
onnxruntime==1.22.1
packaging==25.0
pillow==11.3.0
playsound==1.3.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
pymongo==4.15.1
pyparsing==3.2.3
PyPDF2==3.0.1
pyreadline3==3.5.4
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rsa==4.9.1
runs==1.2.2
setuptools==80.9.0
sniffio==1.3.1
SpeechRecognition==3.14.3
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
websockets==15.0.1
xmod==1.8.1